Added Initial Application Skeleton
This is the first commit for the migration from PHP to Python.
The chosen tool, Django, is more familiar to the developer.
In addition, any data analysis tools are easier to introduce if the
work is done in python.

The current database model will be completely ignored in favor a new
setup. Prior work will be moved over to the new schema.